frontmatter: Fix date
Fix the error:
The page build failed for the `master` branch with the following error:
The value '"\\vspace{-0.8in}sjackman@gmail.com &bull; +1-778-235-0342 &bull; https://sjackman.ca"'
was passed to a date-related filter that expects valid dates in `/_layouts/default.html` or one of its layouts.
For more information, see https://help.github.com/en/articles/page-build-failed-date-is-not-a-valid-datetime.